Add files via upload
The project comes with the technique of OCR (Optical Character Recognition) which
includes various research sides of computer science. The project is to take a picture of a
character and process it up to recognize the image of that character like a human brain
recognize the various digits. The project contains the deep idea of the Image processing
techniques and the big research area of machine learning and the building block of the
machine learning called Neural Network.